+++ source/databricks/aggregation-jobs/aggregation_trigger.py
@@ -20,8 +20,8 @@ sys.path.append(r'/opt/conda/lib/python3.8/site-packages')
 import configargparse
 from geh_stream.shared.data_exporter import export_to_csv
 from geh_stream.aggregation_utils.trigger_base_arguments import trigger_base_arguments
+from geh_stream.shared.data_loader import initialize_spark
 from geh_stream.aggregation_utils.aggregators import \
-    initialize_spark, \
     get_time_series_dataframe, \
     aggregate_net_exchange_per_ga, \
     aggregate_net_exchange_per_neighbour_ga, \
