+++ source/databricks/geh_stream/schemas/output/calculate_daily_subscription_price_schema.py
@@ -21,11 +21,11 @@ calculate_daily_subscription_price_schema = StructType([
       StructField(Colname.charge_id, StringType(), False),
       StructField(Colname.charge_type, StringType(), False),
       StructField(Colname.charge_owner, StringType(), False),
-      StructField(Colname.charge_price, DecimalType(18, 6), False),
+      StructField(Colname.charge_price, DecimalType(18, 8), False),
       StructField(Colname.time, TimestampType(), False),
-      StructField(Colname.price_per_day, DecimalType(18, 6), False),
+      StructField(Colname.price_per_day, DecimalType(18, 8), False),
       StructField(Colname.charge_count, IntegerType(), False),
-      StructField(Colname.total_daily_charge_price, DecimalType(18, 6), False),
+      StructField(Colname.total_daily_charge_price, DecimalType(18, 8), False),
       StructField(Colname.metering_point_type, StringType(), False),
       StructField(Colname.settlement_method, StringType(), False),
       StructField(Colname.grid_area, StringType(), False),
