+++ source/databricks/geh_stream/shared/filters.py
@@ -19,11 +19,13 @@ from typing import List
 
 
 def filter_on_date(df: DataFrame, period: Period) -> DataFrame:
-    return df.filter(col(Colname.time) >= period.from_date).filter(col(Colname.time) < period.to_date)
+    return df
+   # return df.filter(col(Colname.time) >= period.from_date).filter(col(Colname.time) < period.to_date)
 
 
 def filter_on_period(df: DataFrame, period: Period) -> DataFrame:
-    return df.filter((col(Colname.from_date) < period.to_date) & (col(Colname.to_date) > period.from_date))
+    return df
+    # return df.filter((col(Colname.from_date) < period.to_date) & (col(Colname.to_date) > period.from_date))
 
 
 def filter_on_grid_areas(df: DataFrame, grid_area_col: str, grid_areas: List[str]) -> DataFrame:
