+++ source/databricks/geh_stream/wholesale_utils/calculators/tariff_calculators.py
@@ -14,7 +14,8 @@
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import col, sum, count
 from geh_stream.codelists import Colname, ChargeType
-
+from pyspark.sql.types import DecimalType
+from geh_stream.schemas.output import calculate_tariff_price_per_ga_co_es_schema
 
 total_quantity = "total_quantity"
 charge_count = "charge_count"
@@ -59,7 +60,7 @@ def calculate_tariff_price_per_ga_co_es(tariffs: DataFrame) -> DataFrame:
         Colname.settlement_method,
         Colname.charge_key
     ], "inner") \
-    .withColumn("total_amount", col(Colname.charge_price) * col(total_quantity)) \
+    .withColumn("total_amount", col(Colname.charge_price) * col(total_quantity))) \
     .orderBy([Colname.charge_key, Colname.grid_area, Colname.energy_supplier_id, Colname.time, Colname.metering_point_type, Colname.settlement_method])
 
-    return df
+    return spark.createDataFrame(df.rdd, calculate_tariff_price_per_ga_co_es_schema)
